Refactor to use functional adaptor instead of abstract class
