+++ packages/rehype-mathjax/renderer/adaptor.browser.js
@@ -1,0 +1,1 @@
+module.exports = require('mathjax-full/js/adaptors/browserAdaptor').browserAdaptor
