+++ packages/rehype-mathjax/renderer/adaptor.js
@@ -1,0 +1,9 @@
+const JsDom = require('jsdom').JSDOM
+const adaptor = require('mathjax-full/js/adaptors/jsdomAdaptor').jsdomAdaptor
+const register = require('mathjax-full/js/handlers/html').RegisterHTMLHandler
+
+module.exports = createAdaptor
+
+function createAdaptor() {
+  return adaptor(JsDom)
+}
